fix:git conflict해결

+++ src/main/java/jeje/work/aeatbe/repository/ProductFreeFromRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductFreeFromRepository extends JpaRepository<ProductFreeFrom, Integer> {
+public interface ProductFreeFromRepository extends JpaRepository<ProductFreeFrom, Long> {
 
 }
